refactor(karta)!: move the jq engine under internal
pkg/jq becomes internal/jq: the evaluator and its validation are
plumbing behind the consumer surfaces (pkg/karta, pkg/resource,
pkg/tree, the v1alpha1 validator) and could previously be imported
directly, bypassing every capability check the front door enforces.
all in-repo importers move with it; nested modules under the
github.com/run-ai/karta path prefix can still reach it, external
importers cannot.

BREAKING CHANGE: github.com/run-ai/karta/pkg/jq and pkg/jq/execution
are no longer importable outside the repository. consumers use
pkg/karta (or pkg/resource) instead.
