+++ pkg/resource/accessor.go
@@ -15,8 +15,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/run-ai/karta/internal/jq/execution"
 	"github.com/run-ai/karta/pkg/api/runai/v1alpha1"
-	"github.com/run-ai/karta/pkg/jq/execution"
 )
 
 // DefinitionNotFoundError represents an error when a requested definition is not found
