+++ skills/add-workload-type/reference/troubleshooting.md
@@ -5,7 +5,7 @@
 
 Match the error text to a row and apply the fix. Messages come from the
 validator (`pkg/api/runai/v1alpha1/validation.go`), the jq validator
-(`pkg/jq/validation.go`), or the Go accessor API at runtime. The prose version
+(`internal/jq/validation.go`), or the Go accessor API at runtime. The prose version
 is `docs/Troubleshooting.md`.
 
 ## Structure validation errors
