+++ test/blackbox/suspend_resume_test.go
@@ -10,8 +10,8 @@ import (
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/run-ai/karta/internal/jq/execution"
 	"github.com/run-ai/karta/pkg/api/runai/v1alpha1"
-	"github.com/run-ai/karta/pkg/jq/execution"
 	"github.com/run-ai/karta/pkg/resource"
 	"github.com/run-ai/karta/test/types"
 )
